Fixed style errors by moving App page
