+++ src/renderer/pages/CreateNewProjectPage.vue
@@ -35,7 +35,7 @@ export default defineComponent({
 });
 </script>
 
-<style lang="scss">
+<style lang="scss" scoped>
   @import "../styles/main.scss";
   @import "../styles/pages/create-new-project.scss";
 </style>
