+++ src/renderer/pages/OnOpenPage.vue
@@ -45,7 +45,7 @@ export default defineComponent({
 
 </script>
 
-<style lang="scss">
+<style lang="scss" scoped>
   @import "../styles/main.scss";
   @import "../styles/pages/on-open-page.scss";
 </style>
