+++ src/renderer/pages/VideoEditorPage.vue
@@ -73,7 +73,7 @@ export default defineComponent({
 });
 </script>
 
-<style lang="scss">
+<style lang="scss" scoped>
   @import "../styles/main.scss";
   @import "../styles/pages/video-editor.scss";
 </style>
