+++ src/renderer/styles/pages/on-open-page.scss
@@ -1,7 +1,8 @@
 
 
 .horizontal-options-menu {
-
+    height: 45%;
+    
     &>div {
         width: 200px;
         margin-top: 10px;
